add placeholder script to register issues to github
